feat(dashboard): merge the weather and weather-meteo-france widgets into one
