+++ front/src/config/i18n/fr.json
@@ -288,26 +288,14 @@
       "gauge": "Jauge",
       "energy-consumption": "Consommation Énergétique",
       "voice-assistant": "Assistant vocal",
-      "link": "Lien",
-      "weather-meteo-france": "Météo France"
+      "link": "Lien"
     },
     "boxes": {
       "column": "Colonne {{index}}",
       "deleteButton": "Supprimer",
       "weather": {
         "editHouseLabel": "Sélectionnez une maison. J'utiliserai les latitude/longitude pour obtenir la météo.",
-        "houseHasNoCoordinates": "Les coordonnées de votre maison ne sont pas définies. Accédez aux paramètres Gladys pour définir la position de votre maison.",
-        "serviceNotConfigured": "Le service OpenWeather n'est pas configuré. Veuillez accéder à l'onglet «Intégrations» et configurer le service OpenWeather.",
-        "requestToThirdPartyFailed": "La requête à l'API OpenWeather a échoué. Votre instance Gladys est-elle connectée à Internet? Veuillez vous rendre dans le panneau de configuration de OpenWeather pour résoudre ce problème.",
-        "clickHere": "Cliquez ici pour accéder au panneau de configuration de OpenWeather.",
-        "unknownError": "Nous ne pouvons pas obtenir la météo pour cette maison. Avez-vous défini une maison pour cette box ?",
-        "editModeLabel": "Mode d'affichage :",
-        "displayModes": {
-          "advancedWeather": "Humidité et vitesse du vent",
-          "hourlyForecast": "Prévisions pour les 24 prochaines heures",
-          "dailyForecast": "Prévisions pour les 5 prochains jours"
-        },
-        "minMaxDegreeValue": "{{min}}°/{{max}}°"
+        "editModeLabel": "Mode d'affichage :"
       },
       "weatherMeteoFrance": {
         "sourceLabel": "Source des prévisions",
